Immediately stop on an invalid step regex.
If an invalid regex is stored in the STEP_REGISTRY,
depending on the iteration order of the registry,
when the regex is searched (regexes are not stored
compiled) it generates an error or ends up with
a step getting skipped.

The behavior now is to stop as soon as a regex
that won't compile is found.