parses out '@' from CLI and test for cascading tag
CLI now accepts @ for the tag argument without failing.  This was more
because I'm lazy and didn't want to update all our CI runners that I had
set up for my pull request, but also because I think it's nice to have
the option of doing both.  Many people I've talked to consider the tag
to include the '@'

I also added a very basic test that covers the cascading of tags from
feature to their child scenarios.  Hope it's helpful, and I really hope
I have the chance to get back to submitting on it

[272]